PhoneNumber Lib
**Description**
PhoneNumber class is a helper library similar to Time or DateTime
libraries that allows the caller to generate a formattable phone number.
The provided primitive string is digested into the various parts of a
valid phone number and can be accessed with formatter strings. This
ensures that the string has a consistent representation across the
platform by leveraging either a default format or a custom format.

Leveraging the PhoneNumber class is the PhoneNumberValidator and
PhoneNumberType. The validator uses the PhoneNumber class to validate an
incoming string value for an ActiveRecord model to ensure that it is a
valid phone number. The type provides serialization to store the data as
a preformatted string, and access the data model field as a PhoneNumber.

Together these libraries work together to provide data consistency
across the platform.