Application Service
**Description**
This update adds the application service which is a defined service that
provides an interface that allows offloading excessive work from the
controller while keeping the controller tidy and not required to know
implementation details of the downstream code. This eliminates the
controller from having excessive rescue blocks and encapsulates
functional logic beyond model interactions.